twitter_api/requirements.txt
Django
djangorestframework
tweepy
coreapi
markdown